Control.Monad.Class.MonadAsync

Description

async API compatible with both IO and IOSim.

Documentation

class (MonadSTM m, MonadThread m) => MonadAsync m where

Minimal complete definition

async, asyncBound, asyncOn, asyncThreadId, cancel, cancelWith, asyncWithUnmask, asyncOnWithUnmask, waitCatchSTM, pollSTM

Associated Types

type Async m = (async :: Type -> Type) | async -> m

An asynchronous action

See Async.

Methods

async :: m a -> m (Async m a)

See async.

asyncBound :: m a -> m (Async m a)

See asyncBound.

asyncOn :: Int -> m a -> m (Async m a)

See asyncOn.

asyncThreadId :: Async m a -> ThreadId m

See asyncThreadId.

withAsync :: m a -> (Async m a -> m b) -> m b

See withAsync.

default withAsync :: MonadMask m => m a -> (Async m a -> m b) -> m b

withAsyncBound :: m a -> (Async m a -> m b) -> m b

See withAsyncBound.

default withAsyncBound :: MonadMask m => m a -> (Async m a -> m b) -> m b

withAsyncOn :: Int -> m a -> (Async m a -> m b) -> m b

See withAsyncOn.

default withAsyncOn :: MonadMask m => Int -> m a -> (Async m a -> m b) -> m b

waitSTM :: Async m a -> STM m a

See waitSTM.

default waitSTM :: MonadThrow (STM m) => Async m a -> STM m a

pollSTM :: Async m a -> STM m (Maybe (Either SomeException a))

See pollSTM.

waitCatchSTM :: Async m a -> STM m (Either SomeException a)

See waitCatchSTM.

waitAnySTM :: [Async m a] -> STM m (Async m a, a)

See waitAnySTM.

default waitAnySTM :: MonadThrow (STM m) => [Async m a] -> STM m (Async m a, a)

waitAnyCatchSTM :: [Async m a] -> STM m (Async m a, Either SomeException a)

See waitAnyCatchSTM.

waitEitherSTM :: Async m a -> Async m b -> STM m (Either a b)

See waitEitherSTM.

default waitEitherSTM :: MonadThrow (STM m) => Async m a -> Async m b -> STM m (Either a b)

waitEitherSTM_ :: Async m a -> Async m b -> STM m ()

See waitEitherSTM_.

default waitEitherSTM_ :: MonadThrow (STM m) => Async m a -> Async m b -> STM m ()

waitEitherCatchSTM :: Async m a -> Async m b -> STM m (Either (Either SomeException a) (Either SomeException b))

See waitEitherCatchSTM.

waitBothSTM :: Async m a -> Async m b -> STM m (a, b)

See waitBothSTM.

default waitBothSTM :: MonadThrow (STM m) => Async m a -> Async m b -> STM m (a, b)

wait :: Async m a -> m a

See wait.

poll :: Async m a -> m (Maybe (Either SomeException a))

See poll.

waitCatch :: Async m a -> m (Either SomeException a)

See waitCatch.

cancel :: Async m a -> m ()

See cancel.

cancelWith :: Exception e => Async m a -> e -> m ()

See cancelWith.

uninterruptibleCancel :: Async m a -> m ()

See uninterruptibleCancel.

default uninterruptibleCancel :: MonadMask m => Async m a -> m ()

waitAny :: [Async m a] -> m (Async m a, a)

See waitAny.

waitAnyCatch :: [Async m a] -> m (Async m a, Either SomeException a)

See waitAnyCatch.

waitAnyCancel :: [Async m a] -> m (Async m a, a)

See waitAnyCancel.

default waitAnyCancel :: MonadThrow m => [Async m a] -> m (Async m a, a)

waitAnyCatchCancel :: [Async m a] -> m (Async m a, Either SomeException a)

See waitAnyCatchCancel.

default waitAnyCatchCancel :: MonadThrow m => [Async m a] -> m (Async m a, Either SomeException a)

waitEither :: Async m a -> Async m b -> m (Either a b)

See waitEither.

waitEitherCatch :: Async m a -> Async m b -> m (Either (Either SomeException a) (Either SomeException b))

Note, IO-based implementations should override the default implementation. See the async package implementation and comments. http://hackage.haskell.org/package/async-2.2.1/docs/src/Control.Concurrent.Async.html#waitEitherCatch

See waitEitherCatch.

waitEitherCancel :: Async m a -> Async m b -> m (Either a b)

See waitEitherCancel.

default waitEitherCancel :: MonadThrow m => Async m a -> Async m b -> m (Either a b)

waitEitherCatchCancel :: Async m a -> Async m b -> m (Either (Either SomeException a) (Either SomeException b))

See waitEitherCatchCancel.

default waitEitherCatchCancel :: MonadThrow m => Async m a -> Async m b -> m (Either (Either SomeException a) (Either SomeException b))

waitEither_ :: Async m a -> Async m b -> m ()

See waitEither_.

waitBoth :: Async m a -> Async m b -> m (a, b)

See waitBoth.

race :: m a -> m b -> m (Either a b)

See race.

race_ :: m a -> m b -> m ()

See race_.

concurrently :: m a -> m b -> m (a, b)

See concurrently.

concurrently_ :: m a -> m b -> m ()

See concurrently_.

asyncWithUnmask :: ((forall b. m b -> m b) -> m a) -> m (Async m a)

See concurrently_.

asyncOnWithUnmask :: Int -> ((forall b. m b -> m b) -> m a) -> m (Async m a)

See asyncOnWithUnmask.

withAsyncWithUnmask :: ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

See withAsyncWithUnmask.

default withAsyncWithUnmask :: MonadMask m => ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

withAsyncOnWithUnmask :: Int -> ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

See withAsyncOnWithUnmask.

default withAsyncOnWithUnmask :: MonadMask m => Int -> ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

compareAsyncs :: Async m a -> Async m b -> Ordering

See compareAsyncs.

default compareAsyncs :: Ord (ThreadId m) => Async m a -> Async m b -> Ordering

data AsyncCancelled

Constructors

AsyncCancelled

data ExceptionInLinkedThread

Exception from child thread re-raised in parent thread

We record the thread ID of the child thread as a String. This avoids an m parameter in the type, which is important: ExceptionInLinkedThread must be an instance of Exception, requiring it to be Typeable; if m appeared in the type, we would require m to be Typeable, which does not work with with the simulator, as it would require a Typeable constraint on the s parameter of IOSim.

Constructors

ExceptionInLinkedThread String SomeException

link :: (MonadAsync m, MonadFork m, MonadMask m) => Async m a -> m ()

Like link.

linkOnly :: forall m a. (MonadAsync m, MonadFork m, MonadMask m) => (SomeException -> Bool) -> Async m a -> m ()

Like linkOnly.

link2 :: (MonadAsync m, MonadFork m, MonadMask m) => Async m a -> Async m b -> m ()

Like link2.

link2Only :: (MonadAsync m, MonadFork m, MonadMask m) => (SomeException -> Bool) -> Async m a -> Async m b -> m ()

Like link2Only.

mapConcurrently :: (Traversable t, MonadAsync m) => (a -> m b) -> t a -> m (t b)

See mapConcurrently.

forConcurrently :: (Traversable t, MonadAsync m) => t a -> (a -> m b) -> m (t b)

See forConcurrently.

mapConcurrently_ :: (Foldable f, MonadAsync m) => (a -> m b) -> f a -> m ()

See mapConcurrently_.

forConcurrently_ :: (Foldable f, MonadAsync m) => f a -> (a -> m b) -> m ()

See forConcurrently_.

replicateConcurrently :: MonadAsync m => Int -> m a -> m [a]

See replicateConcurrently.

replicateConcurrently_ :: MonadAsync m => Int -> m a -> m ()

See replicateConcurrently_.

newtype Concurrently m a

Similar to Concurrently but which works for any MonadAsync instance.

Constructors

Concurrently

Fields

runConcurrently :: m a