Control.Monad.Class.MonadSTM

Documentation

class (Monad m, Alternative (STM m), MonadPlus (STM m)) ⇒ MonadSTM m where

Minimal complete definition

atomically, newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, writeTQueue, isEmptyTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue

Associated Types

type STM m = (stm ∷ Type → Type) | stm → m

type TVar m ∷ Type → Type

type TMVar m ∷ Type → Type

type TQueue m ∷ Type → Type

type TBQueue m ∷ Type → Type

Methods

atomically ∷ HasCallStack ⇒ STM m a → m a

newTVar ∷ a → STM m (TVar m a)

readTVar ∷ TVar m a → STM m a

writeTVar ∷ TVar m a → a → STM m ()

retry ∷ STM m a

orElse ∷ STM m a → STM m a → STM m a

modifyTVar ∷ TVar m a → (a → a) → STM m ()

modifyTVar' ∷ TVar m a → (a → a) → STM m ()

stateTVar ∷ TVar m s → (s → (a, s)) → STM m a

@since io-classes-0.2.0.0

swapTVar ∷ TVar m a → a → STM m a

check ∷ Bool → STM m ()

newTMVar ∷ a → STM m (TMVar m a)

newEmptyTMVar ∷ STM m (TMVar m a)

takeTMVar ∷ TMVar m a → STM m a

tryTakeTMVar ∷ TMVar m a → STM m (Maybe a)

putTMVar ∷ TMVar m a → a → STM m ()

tryPutTMVar ∷ TMVar m a → a → STM m Bool

readTMVar ∷ TMVar m a → STM m a

tryReadTMVar ∷ TMVar m a → STM m (Maybe a)

swapTMVar ∷ TMVar m a → a → STM m a

isEmptyTMVar ∷ TMVar m a → STM m Bool

newTQueue ∷ STM m (TQueue m a)

readTQueue ∷ TQueue m a → STM m a

tryReadTQueue ∷ TQueue m a → STM m (Maybe a)

peekTQueue ∷ TQueue m a → STM m a

tryPeekTQueue ∷ TQueue m a → STM m (Maybe a)

writeTQueue ∷ TQueue m a → a → STM m ()

isEmptyTQueue ∷ TQueue m a → STM m Bool

newTBQueue ∷ Natural → STM m (TBQueue m a)

readTBQueue ∷ TBQueue m a → STM m a

tryReadTBQueue ∷ TBQueue m a → STM m (Maybe a)

peekTBQueue ∷ TBQueue m a → STM m a

tryPeekTBQueue ∷ TBQueue m a → STM m (Maybe a)

flushTBQueue ∷ TBQueue m a → STM m [a]

writeTBQueue ∷ TBQueue m a → a → STM m ()

lengthTBQueue ∷ TBQueue m a → STM m Natural

Since: 0.2.0.0

isEmptyTBQueue ∷ TBQueue m a → STM m Bool

isFullTBQueue ∷ TBQueue m a → STM m Bool

newTVarIO ∷ a → m (TVar m a)

readTVarIO ∷ TVar m a → m a

newTMVarIO ∷ a → m (TMVar m a)

newEmptyTMVarIO ∷ m (TMVar m a)

newTQueueIO ∷ m (TQueue m a)

newTBQueueIO ∷ Natural → m (TBQueue m a)

class MonadSTM m ⇒ MonadLabelledSTM m where

Labelled TVars, TMVars, TQueues and TBQueues.

Minimal complete definition

labelTVar, labelTMVar, labelTQueue, labelTBQueue

Methods

labelTVar ∷ TVar m a → String → STM m ()

labelTMVar ∷ TMVar m a → String → STM m ()

labelTQueue ∷ TQueue m a → String → STM m ()

labelTBQueue ∷ TBQueue m a → String → STM m ()

labelTVarIO ∷ TVar m a → String → m ()

default labelTVarIO ∷ TVar m a → String → m ()

labelTMVarIO ∷ TMVar m a → String → m ()

default labelTMVarIO ∷ TMVar m a → String → m ()

labelTQueueIO ∷ TQueue m a → String → m ()

default labelTQueueIO ∷ TQueue m a → String → m ()

labelTBQueueIO ∷ TBQueue m a → String → m ()

default labelTBQueueIO ∷ TBQueue m a → String → m ()

class (MonadSTM m, Monad (InspectMonad m)) ⇒ MonadInspectSTM m where

This type class is indented for 'io-sim', where one might want to access TVar in the underlying ST monad.

Associated Types

type InspectMonad m ∷ Type → Type

Methods

inspectTVar ∷ proxy m → TVar m a → InspectMonad m a

inspectTMVar ∷ proxy m → TMVar m a → InspectMonad m (Maybe a)

class MonadInspectSTM m ⇒ MonadTraceSTM m where

MonadTraceSTM allows to trace values of stm variables when stm transaction is commited. This allows to verify invariants when a variable is commited.

Minimal complete definition

traceTVar, traceTQueue, traceTBQueue

Methods

traceTVar

Arguments

∷ Typeable tr

⇒ proxy m

→ TVar m a

→ (Maybe a → a → InspectMonad m tr)
callback which receives initial value or Nothing (if it is a newly created TVar), and the commited value.

→ STM m ()

traceTMVar ∷ Typeable tr ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → STM m ()

default traceTMVar ∷ (TMVar m a ~ TMVarDefault m a, Typeable tr) ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → STM m ()

traceTQueue ∷ Typeable tr ⇒ proxy m → TQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → STM m ()

traceTBQueue ∷ Typeable tr ⇒ proxy m → TBQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → STM m ()

traceTVarIO ∷ Typeable tr ⇒ proxy m → TVar m a → (Maybe a → a → InspectMonad m tr) → m ()

default traceTVarIO ∷ Typeable tr ⇒ proxy m → TVar m a → (Maybe a → a → InspectMonad m tr) → m ()

traceTMVarIO ∷ Typeable tr ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → m ()

default traceTMVarIO ∷ Typeable tr ⇒ proxy m → TMVar m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → m ()

traceTQueueIO ∷ Typeable tr ⇒ proxy m → TQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

default traceTQueueIO ∷ Typeable tr ⇒ proxy m → TQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

traceTBQueueIO ∷ Typeable tr ⇒ proxy m → TBQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

default traceTBQueueIO ∷ Typeable tr ⇒ proxy m → TBQueue m a → (Maybe [a] → [a] → InspectMonad m tr) → m ()

type LazyTVar m = TVar m

Deprecated: Renamed back to TVar

type LazyTMVar m = TMVar m

Deprecated: Renamed back to TMVar

Default TMVar implementation

newtype TMVarDefault m a

Constructors

TMVar (TVar m (Maybe a))

labelTMVarDefault ∷ MonadLabelledSTM m ⇒ TMVarDefault m a → String → STM m ()

traceTMVarDefault ∷ (MonadTraceSTM m, Typeable tr) ⇒ proxy m → TMVarDefault m a → (Maybe (Maybe a) → Maybe a → InspectMonad m tr) → STM m ()

newTMVarDefault ∷ MonadSTM m ⇒ a → STM m (TMVarDefault m a)

newTMVarIODefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newEmptyTMVarDefault ∷ MonadSTM m ⇒ STM m (TMVarDefault m a)

newEmptyTMVarIODefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

takeTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m a

tryTakeTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

putTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m ()

tryPutTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m Bool

readTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m a

tryReadTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

swapTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → a → STM m a

isEmptyTMVarDefault ∷ MonadSTM m ⇒ TMVarDefault m a → STM m Bool

Default TBQueue implementation

data TQueueDefault m a

Constructors

TQueue !(TVar m [a]) !(TVar m [a])

labelTQueueDefault ∷ MonadLabelledSTM m ⇒ TQueueDefault m a → String → STM m ()

newTQueueDefault ∷ MonadSTM m ⇒ STM m (TQueueDefault m a)

readTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m a

tryReadTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

peekTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m a

tryPeekTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

writeTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → a → STM m ()

isEmptyTQueueDefault ∷ MonadSTM m ⇒ TQueueDefault m a → STM m Bool

Default TBQueue implementation

data TBQueueDefault m a

Constructors

TBQueue !(TVar m Natural) !(TVar m [a]) !(TVar m Natural) !(TVar m [a]) !Natural

labelTBQueueDefault ∷ MonadLabelledSTM m ⇒ TBQueueDefault m a → String → STM m ()

newTBQueueDefault ∷ MonadSTM m ⇒ Natural → STM m (TBQueueDefault m a)

readTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryReadTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

peekTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryPeekTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

writeTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → a → STM m ()

isEmptyTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

isFullTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

lengthTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m Natural

flushTBQueueDefault ∷ MonadSTM m ⇒ TBQueueDefault m a → STM m [a]

MonadThrow aliases

throwSTM ∷ (MonadSTM m, MonadThrow (STM m), Exception e) ⇒ e → STM m a

throwIO specialised to stm monad.

catchSTM ∷ (MonadSTM m, MonadCatch (STM m), Exception e) ⇒ STM m a → (e → STM m a) → STM m a

catch speclialized for an stm monad.

Deprecated API

newTVarM ∷ MonadSTM m ⇒ a → m (TVar m a)

Deprecated: Use newTVarIO

newTMVarM ∷ MonadSTM m ⇒ a → m (TMVar m a)

Deprecated: Use newTMVarIO

newTMVarMDefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

Deprecated: Use newTMVarIODefault

newEmptyTMVarM ∷ MonadSTM m ⇒ m (TMVar m a)

Deprecated: Use newEmptyTMVarIO

newEmptyTMVarMDefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

Deprecated: Use newEmptyTMVarIODefault

newtype WrappedSTM (t ∷ Trans) r (m ∷ Type → Type) a

A newtype wrapper for an STM monad for monad transformers.

Constructors

WrappedSTM

Fields

runWrappedSTM ∷ STM m a