Control.Monad.Class.MonadSTM.Internal

Documentation

class (Monad m, Monad (STM m)) => MonadSTM m where

The STM primitives parametrised by a monad m.

Minimal complete definition

atomically, newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, writeTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, flushTQueue, writeTQueue, isEmptyTQueue, unGetTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue, unGetTBQueue, newTSem, waitTSem, signalTSem, signalTSemN, newTChan, newBroadcastTChan, dupTChan, cloneTChan, readTChan, tryReadTChan, peekTChan, tryPeekTChan, writeTChan, unGetTChan, isEmptyTChan

Associated Types

type STM m = (stm :: Type -> Type) | stm -> m

The STM monad.

type TVar m :: Type -> Type

A type of a TVar.

See TVar.

type TMVar m :: Type -> Type

type TQueue m :: Type -> Type

type TBQueue m :: Type -> Type

type TArray m :: Type -> Type -> Type

type TSem m :: Type

type TChan m :: Type -> Type

Methods

atomically :: HasCallStack => STM m a -> m a

Atomically run an STM computation.

See atomically.

newTVar :: a -> STM m (TVar m a)

readTVar :: TVar m a -> STM m a

writeTVar :: TVar m a -> a -> STM m ()

retry :: STM m a

See retry.

orElse :: STM m a -> STM m a -> STM m a

See orElse.

modifyTVar :: TVar m a -> (a -> a) -> STM m ()

modifyTVar' :: TVar m a -> (a -> a) -> STM m ()

stateTVar :: TVar m s -> (s -> (a, s)) -> STM m a

@since io-classes-0.2.0.0

swapTVar :: TVar m a -> a -> STM m a

check :: Bool -> STM m ()

See check.

newTMVar :: a -> STM m (TMVar m a)

newEmptyTMVar :: STM m (TMVar m a)

takeTMVar :: TMVar m a -> STM m a

tryTakeTMVar :: TMVar m a -> STM m (Maybe a)

putTMVar :: TMVar m a -> a -> STM m ()

tryPutTMVar :: TMVar m a -> a -> STM m Bool

readTMVar :: TMVar m a -> STM m a

tryReadTMVar :: TMVar m a -> STM m (Maybe a)

swapTMVar :: TMVar m a -> a -> STM m a

writeTMVar :: TMVar m a -> a -> STM m ()

isEmptyTMVar :: TMVar m a -> STM m Bool

newTQueue :: STM m (TQueue m a)

readTQueue :: TQueue m a -> STM m a

tryReadTQueue :: TQueue m a -> STM m (Maybe a)

peekTQueue :: TQueue m a -> STM m a

tryPeekTQueue :: TQueue m a -> STM m (Maybe a)

flushTQueue :: TQueue m a -> STM m [a]

writeTQueue :: TQueue m a -> a -> STM m ()

isEmptyTQueue :: TQueue m a -> STM m Bool

unGetTQueue :: TQueue m a -> a -> STM m ()

newTBQueue :: Natural -> STM m (TBQueue m a)

readTBQueue :: TBQueue m a -> STM m a

tryReadTBQueue :: TBQueue m a -> STM m (Maybe a)

peekTBQueue :: TBQueue m a -> STM m a

tryPeekTBQueue :: TBQueue m a -> STM m (Maybe a)

flushTBQueue :: TBQueue m a -> STM m [a]

writeTBQueue :: TBQueue m a -> a -> STM m ()

lengthTBQueue :: TBQueue m a -> STM m Natural

Since: 0.2.0.0

isEmptyTBQueue :: TBQueue m a -> STM m Bool

isFullTBQueue :: TBQueue m a -> STM m Bool

unGetTBQueue :: TBQueue m a -> a -> STM m ()

newTSem :: Integer -> STM m (TSem m)

waitTSem :: TSem m -> STM m ()

signalTSem :: TSem m -> STM m ()

signalTSemN :: Natural -> TSem m -> STM m ()

newTChan :: STM m (TChan m a)

newBroadcastTChan :: STM m (TChan m a)

dupTChan :: TChan m a -> STM m (TChan m a)

cloneTChan :: TChan m a -> STM m (TChan m a)

readTChan :: TChan m a -> STM m a

tryReadTChan :: TChan m a -> STM m (Maybe a)

peekTChan :: TChan m a -> STM m a

tryPeekTChan :: TChan m a -> STM m (Maybe a)

writeTChan :: TChan m a -> a -> STM m ()

unGetTChan :: TChan m a -> a -> STM m ()

isEmptyTChan :: TChan m a -> STM m Bool

newTVarIO :: a -> m (TVar m a)

readTVarIO :: TVar m a -> m a

newTMVarIO :: a -> m (TMVar m a)

newEmptyTMVarIO :: m (TMVar m a)

newTQueueIO :: m (TQueue m a)

newTBQueueIO :: Natural -> m (TBQueue m a)

newTChanIO :: m (TChan m a)

newBroadcastTChanIO :: m (TChan m a)

class MonadSTM m => MonadLabelledSTM m where

Labelled TVars & friends.

The IO instances is no-op, the IOSim instance enhances simulation trace. This is very useful when analysing low lever concurrency issues (e.g. deadlocks, livelocks etc).

Minimal complete definition

labelTVar

Methods

labelTVar :: TVar m a -> String -> STM m ()

Name a TVar.

labelTMVar :: TMVar m a -> String -> STM m ()

default labelTMVar :: TMVar m ~ TMVarDefault m => TMVar m a -> String -> STM m ()

labelTQueue :: TQueue m a -> String -> STM m ()

default labelTQueue :: TQueue m ~ TQueueDefault m => TQueue m a -> String -> STM m ()

labelTBQueue :: TBQueue m a -> String -> STM m ()

default labelTBQueue :: TBQueue m ~ TBQueueDefault m => TBQueue m a -> String -> STM m ()

labelTArray :: (Ix i, Show i) => TArray m i e -> String -> STM m ()

default labelTArray :: (TArray m ~ TArrayDefault m, Ix i, Show i) => TArray m i e -> String -> STM m ()

labelTSem :: TSem m -> String -> STM m ()

default labelTSem :: TSem m ~ TSemDefault m => TSem m -> String -> STM m ()

labelTChan :: TChan m a -> String -> STM m ()

default labelTChan :: TChan m ~ TChanDefault m => TChan m a -> String -> STM m ()

labelTVarIO :: TVar m a -> String -> m ()

default labelTVarIO :: TVar m a -> String -> m ()

labelTMVarIO :: TMVar m a -> String -> m ()

default labelTMVarIO :: TMVar m a -> String -> m ()

labelTQueueIO :: TQueue m a -> String -> m ()

default labelTQueueIO :: TQueue m a -> String -> m ()

labelTBQueueIO :: TBQueue m a -> String -> m ()

default labelTBQueueIO :: TBQueue m a -> String -> m ()

labelTArrayIO :: (Ix i, Show i) => TArray m i e -> String -> m ()

default labelTArrayIO :: (Ix i, Show i) => TArray m i e -> String -> m ()

labelTSemIO :: TSem m -> String -> m ()

default labelTSemIO :: TSem m -> String -> m ()

labelTChanIO :: TChan m a -> String -> m ()

default labelTChanIO :: TChan m a -> String -> m ()

class (MonadSTM m, Monad (InspectMonad m)) => MonadInspectSTM m where

This type class is indented for 'io-sim', where one might want to access a TVar in the underlying ST monad.

Associated Types

type InspectMonad m :: Type -> Type

Methods

inspectTVar :: proxy m -> TVar m a -> InspectMonad m a

Return the value of a TVar as an InspectMonad computation.

inspectTVar is useful if the value of a TVar observed by traceTVar contains other TVars.

inspectTMVar :: proxy m -> TMVar m a -> InspectMonad m (Maybe a)

Return the value of a MonadSTM as an InspectMonad computation.

data TraceValue where

A GADT which instructs how to trace the value. The traceDynamic will use dynamic tracing, e.g. "Control.Monad.IOSim.traceM"; while traceString will be traced with EventSay. The IOSims dynamic tracing allows to recover the value from the simulation trace (see "Control.Monad.IOSim.selectTraceEventsDynamic").

Constructors

TraceValue

Fields

:: forall tr. Typeable tr

=> { traceDynamic :: Maybe tr

, traceString :: Maybe String

} -> TraceValue

Bundled Patterns

pattern TraceDynamic :: () => forall tr. Typeable tr => tr -> TraceValue
Use only a dynamic tracer.

pattern TraceString :: String -> TraceValue
Use only string tracing.

pattern DontTrace :: TraceValue
Do not trace the value.

class MonadInspectSTM m => MonadTraceSTM m where

MonadTraceSTM allows to trace values of stm variables when stm transaction is committed. This allows to verify invariants when a variable is committed.

Minimal complete definition

traceTVar, traceTQueue, traceTBQueue

Methods

traceTVar

Arguments

:: proxy m

-> TVar m a

-> (Maybe a -> a -> InspectMonad m TraceValue)
callback which receives initial value or Nothing (if it is a newly created TVar), and the committed value.

-> STM m ()

Construct a trace output out of previous & new value of a TVar. The callback is called whenever an stm transaction which modifies the TVar is committed.

This is supported by IOSim (and IOSimPOR); IO has a trivial instance.

The simplest example is:

traceTVar (Proxy @m) tvar (\_ -> TraceString . show)

Note that the interpretation of TraceValue depends on the monad m itself (see TraceValue).

traceTMVar :: proxy m -> TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> STM m ()

default traceTMVar :: TMVar m a ~ TMVarDefault m a => proxy m -> TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> STM m ()

traceTQueue :: proxy m -> TQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> STM m ()

traceTBQueue :: proxy m -> TBQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> STM m ()

traceTSem :: proxy m -> TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> STM m ()

default traceTSem :: TSem m ~ TSemDefault m => proxy m -> TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> STM m ()

traceTVarIO :: TVar m a -> (Maybe a -> a -> InspectMonad m TraceValue) -> m ()

default traceTVarIO :: TVar m a -> (Maybe a -> a -> InspectMonad m TraceValue) -> m ()

traceTMVarIO :: TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> m ()

default traceTMVarIO :: TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> m ()

traceTQueueIO :: TQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

default traceTQueueIO :: TQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

traceTBQueueIO :: TBQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

default traceTBQueueIO :: TBQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

traceTSemIO :: TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> m ()

default traceTSemIO :: TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> m ()

MonadThrow aliases

throwSTM :: (MonadSTM m, MonadThrow (STM m), Exception e) => e -> STM m a

throwIO specialised to stm monad.

catchSTM :: (MonadSTM m, MonadCatch (STM m), Exception e) => STM m a -> (e -> STM m a) -> STM m a

catch specialized for an stm monad.

Default implementations

The default implementations are based on a TVar defined in the class. They are tailored towards IOSim rather than instances which would like to derive from IO or monad transformers.

newtype TMVarDefault m a

Constructors

TMVar (TVar m (Maybe a))

newTMVarDefault :: MonadSTM m => a -> STM m (TMVarDefault m a)

newEmptyTMVarDefault :: MonadSTM m => STM m (TMVarDefault m a)

takeTMVarDefault :: MonadSTM m => TMVarDefault m a -> STM m a

tryTakeTMVarDefault :: MonadSTM m => TMVarDefault m a -> STM m (Maybe a)

putTMVarDefault :: MonadSTM m => TMVarDefault m a -> a -> STM m ()

tryPutTMVarDefault :: MonadSTM m => TMVarDefault m a -> a -> STM m Bool

readTMVarDefault :: MonadSTM m => TMVarDefault m a -> STM m a

tryReadTMVarDefault :: MonadSTM m => TMVarDefault m a -> STM m (Maybe a)

swapTMVarDefault :: MonadSTM m => TMVarDefault m a -> a -> STM m a

writeTMVarDefault :: MonadSTM m => TMVarDefault m a -> a -> STM m ()

isEmptyTMVarDefault :: MonadSTM m => TMVarDefault m a -> STM m Bool

labelTMVarDefault :: MonadLabelledSTM m => TMVarDefault m a -> String -> STM m ()

traceTMVarDefault :: MonadTraceSTM m => proxy m -> TMVarDefault m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> STM m ()

Default MonadSTM implementation

data TQueueDefault m a

Constructors

TQueue !(TVar m [a]) !(TVar m [a])

newTQueueDefault :: MonadSTM m => STM m (TQueueDefault m a)

writeTQueueDefault :: MonadSTM m => TQueueDefault m a -> a -> STM m ()

readTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m a

tryReadTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m (Maybe a)

isEmptyTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m Bool

peekTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m a

tryPeekTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m (Maybe a)

flushTQueueDefault :: MonadSTM m => TQueueDefault m a -> STM m [a]

unGetTQueueDefault :: MonadSTM m => TQueueDefault m a -> a -> STM m ()

labelTQueueDefault :: MonadLabelledSTM m => TQueueDefault m a -> String -> STM m ()

Default MonadSTM implementation

data TBQueueDefault m a

Constructors

TBQueue !(TVar m Natural) !(TVar m [a]) !(TVar m Natural) !(TVar m [a]) !Natural

newTBQueueDefault :: MonadSTM m => Natural -> STM m (TBQueueDefault m a)

writeTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> a -> STM m ()

readTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m a

tryReadTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m (Maybe a)

peekTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m a

tryPeekTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m (Maybe a)

isEmptyTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m Bool

isFullTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m Bool

lengthTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m Natural

flushTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> STM m [a]

unGetTBQueueDefault :: MonadSTM m => TBQueueDefault m a -> a -> STM m ()

labelTBQueueDefault :: MonadLabelledSTM m => TBQueueDefault m a -> String -> STM m ()

Default MonadSTM implementation

data TArrayDefault m i e

Default implementation of MonadSTM.

Constructors

TArray (Array i (TVar m e))

Default MonadSTM implementation

newtype TSemDefault m

Constructors

TSem (TVar m Integer)

newTSemDefault :: MonadSTM m => Integer -> STM m (TSemDefault m)

waitTSemDefault :: MonadSTM m => TSemDefault m -> STM m ()

signalTSemDefault :: MonadSTM m => TSemDefault m -> STM m ()

signalTSemNDefault :: MonadSTM m => Natural -> TSemDefault m -> STM m ()

labelTSemDefault :: MonadLabelledSTM m => TSemDefault m -> String -> STM m ()

Default MonadSTM implementation

data TChanDefault m a

Constructors

TChan (TVar m (TVarList m a)) (TVar m (TVarList m a))

newTChanDefault :: MonadSTM m => STM m (TChanDefault m a)

newBroadcastTChanDefault :: MonadSTM m => STM m (TChanDefault m a)

writeTChanDefault :: MonadSTM m => TChanDefault m a -> a -> STM m ()

readTChanDefault :: MonadSTM m => TChanDefault m a -> STM m a

tryReadTChanDefault :: MonadSTM m => TChanDefault m a -> STM m (Maybe a)

peekTChanDefault :: MonadSTM m => TChanDefault m a -> STM m a

tryPeekTChanDefault :: MonadSTM m => TChanDefault m a -> STM m (Maybe a)

dupTChanDefault :: MonadSTM m => TChanDefault m a -> STM m (TChanDefault m a)

unGetTChanDefault :: MonadSTM m => TChanDefault m a -> a -> STM m ()

isEmptyTChanDefault :: MonadSTM m => TChanDefault m a -> STM m Bool

cloneTChanDefault :: MonadSTM m => TChanDefault m a -> STM m (TChanDefault m a)

labelTChanDefault :: MonadLabelledSTM m => TChanDefault m a -> String -> STM m ()