Control.Monad.Class.MonadTimer

Description

Provides classes to handle delays and timeouts which generalised base API to both IO and IOSim.

Documentation

class Monad m => MonadDelay m where

A typeclass to delay current thread.

Methods

threadDelay :: Int -> m ()

Suspends the current thread for a given number of microseconds (GHC only).

See threadDelay.

class (MonadDelay m, MonadSTM m) => MonadTimer m where

A typeclass providing utilities for timeouts.

Methods

registerDelay :: Int -> m (TVar m Bool)

See registerDelay.

timeout :: Int -> m a -> m (Maybe a)

See timeout.