Control.Monad.Class.MonadAsync

Documentation

class (MonadSTM m, MonadThread m) ⇒ MonadAsync m where

Minimal complete definition

async, asyncThreadId, cancel, cancelWith, asyncWithUnmask, waitCatchSTM, pollSTM

Associated Types

type Async m = (async ∷ Type → Type) | async → m

An asynchronous action

Methods

async ∷ m a → m (Async m a)

asyncThreadId ∷ Async m a → ThreadId m

withAsync ∷ m a → (Async m a → m b) → m b

default withAsync ∷ MonadMask m ⇒ m a → (Async m a → m b) → m b

waitSTM ∷ Async m a → STM m a

default waitSTM ∷ MonadThrow (STM m) ⇒ Async m a → STM m a

pollSTM ∷ Async m a → STM m (Maybe (Either SomeException a))

waitCatchSTM ∷ Async m a → STM m (Either SomeException a)

waitAnySTM ∷ [Async m a] → STM m (Async m a, a)

default waitAnySTM ∷ MonadThrow (STM m) ⇒ [Async m a] → STM m (Async m a, a)

waitAnyCatchSTM ∷ [Async m a] → STM m (Async m a, Either SomeException a)

waitEitherSTM ∷ Async m a → Async m b → STM m (Either a b)

default waitEitherSTM ∷ MonadThrow (STM m) ⇒ Async m a → Async m b → STM m (Either a b)

waitEitherSTM_ ∷ Async m a → Async m b → STM m ()

default waitEitherSTM_ ∷ MonadThrow (STM m) ⇒ Async m a → Async m b → STM m ()

waitEitherCatchSTM ∷ Async m a → Async m b → STM m (Either (Either SomeException a) (Either SomeException b))

waitBothSTM ∷ Async m a → Async m b → STM m (a, b)

default waitBothSTM ∷ MonadThrow (STM m) ⇒ Async m a → Async m b → STM m (a, b)

wait ∷ Async m a → m a

poll ∷ Async m a → m (Maybe (Either SomeException a))

waitCatch ∷ Async m a → m (Either SomeException a)

cancel ∷ Async m a → m ()

cancelWith ∷ Exception e ⇒ Async m a → e → m ()

uninterruptibleCancel ∷ Async m a → m ()

default uninterruptibleCancel ∷ MonadMask m ⇒ Async m a → m ()

waitAny ∷ [Async m a] → m (Async m a, a)

waitAnyCatch ∷ [Async m a] → m (Async m a, Either SomeException a)

waitAnyCancel ∷ [Async m a] → m (Async m a, a)

default waitAnyCancel ∷ MonadThrow m ⇒ [Async m a] → m (Async m a, a)

waitAnyCatchCancel ∷ [Async m a] → m (Async m a, Either SomeException a)

default waitAnyCatchCancel ∷ MonadThrow m ⇒ [Async m a] → m (Async m a, Either SomeException a)

waitEither ∷ Async m a → Async m b → m (Either a b)

waitEitherCatch ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

Note, IO-based implementations should override the default implementation. See the async package implementation and comments. http://hackage.haskell.org/package/async-2.2.1/docs/src/Control.Concurrent.Async.html#waitEitherCatch

waitEitherCancel ∷ Async m a → Async m b → m (Either a b)

default waitEitherCancel ∷ MonadThrow m ⇒ Async m a → Async m b → m (Either a b)

waitEitherCatchCancel ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

default waitEitherCatchCancel ∷ MonadThrow m ⇒ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

waitEither_ ∷ Async m a → Async m b → m ()

waitBoth ∷ Async m a → Async m b → m (a, b)

race ∷ m a → m b → m (Either a b)

race_ ∷ m a → m b → m ()

concurrently ∷ m a → m b → m (a, b)

concurrently_ ∷ m a → m b → m ()

asyncWithUnmask ∷ ((∀ b. m b → m b) → m a) → m (Async m a)

data AsyncCancelled

Constructors

AsyncCancelled

data ExceptionInLinkedThread

Exception from child thread re-raised in parent thread

We record the thread ID of the child thread as a String. This avoids an m parameter in the type, which is important: ExceptionInLinkedThread must be an instance of Exception, requiring it to be Typeable; if m appeared in the type, we would require m to be Typeable, which does not work with with the simulator, as it would require a Typeable constraint on the s parameter of IOSim.

Constructors

ExceptionInLinkedThread String SomeException

link ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ Async m a → m ()

linkTo ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ ThreadId m → Async m a → m ()

Generalizion of link that links an async to an arbitrary thread.

linkOnly ∷ ∀ m a. (MonadAsync m, MonadFork m, MonadMask m) ⇒ (SomeException → Bool) → Async m a → m ()

linkToOnly ∷ ∀ m a. (MonadAsync m, MonadFork m, MonadMask m) ⇒ ThreadId m → (SomeException → Bool) → Async m a → m ()

mapConcurrently ∷ (Traversable t, MonadAsync m) ⇒ (a → m b) → t a → m (t b)

forConcurrently ∷ (Traversable t, MonadAsync m) ⇒ t a → (a → m b) → m (t b)

mapConcurrently_ ∷ (Foldable f, MonadAsync m) ⇒ (a → m b) → f a → m ()

forConcurrently_ ∷ (Foldable f, MonadAsync m) ⇒ f a → (a → m b) → m ()

replicateConcurrently ∷ MonadAsync m ⇒ Int → m a → m [a]

replicateConcurrently_ ∷ MonadAsync m ⇒ Int → m a → m ()

newtype Concurrently m a

Similar to Concurrently but which works for any MonadAsync instance.

Constructors

Concurrently

Fields

runConcurrently ∷ m a