Control.Monad.Class.MonadEventlog

Documentation

class Monad m => MonadEventlog m where

Methods

traceEventIO :: String -> m ()

Emits a message to the eventlog, if eventlog profiling is available and enabled at runtime.

traceMarkerIO :: String -> m ()

Emits a marker to the eventlog, if eventlog profiling is available and enabled at runtime.

The String is the name of the marker. The name is just used in the profiling tools to help you keep clear which marker is which.