Control.Monad.Class.MonadTime

Description

time and base time API compatible with both IO and IOSim.

Documentation

class Monad m => MonadTime m where

Methods

getCurrentTime :: m UTCTime

Wall clock time.

class Monad m => MonadMonotonicTimeNSec m where

Methods

getMonotonicTimeNSec :: m Word64

Time in a monotonic clock, with high precision. The epoch for this clock is arbitrary and does not correspond to any wall clock or calendar.

The time is measured in nano seconds as does getMonotonicTimeNSec from "base".

NominalTime and its action on UTCTime

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.